% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt-powerline.R
\name{make_prompt_powerline}
\alias{make_prompt_powerline}
\title{This is a powerline-like prompt}
\usage{
make_prompt_powerline(
  color1 = "#404040",
  color2 = "#0055d4",
  color3 = "gold4",
  color4 = "#7d26cd"
)
}
\arguments{
\item{color1, color2, color3, color4}{Background colors for the sections:
allocated memory, working directory, devtools loaded package(s),
git info.}
}
\value{
`make_prompt_powerline()` returns a function that you can
use with [set_prompt()].
}
\description{
It is inspired by the <https://github.com/powerline/powerline>
project.
}
