% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{prompt}
\alias{prompt}
\alias{prompt-package}
\title{Dynamic R Prompt}
\description{
Set the R prompt dynamically, from a function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gaborcsardi/prompt}
  \item Report bugs at \url{https://github.com/gaborcsardi/prompt/issues}
}

}
