% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt-devtools.R
\name{prompt_devtools}
\alias{prompt_devtools}
\alias{devtools_packages}
\title{Example prompt that shows the package being developed with devtools}
\usage{
prompt_devtools(...)

devtools_packages()
}
\arguments{
\item{...}{Ignored.}
}
\value{
\code{prompt_devtools()} returns the prompt string.

\code{devtools_packages()} returns the packages loaded by
devtools/pkgload.
}
\description{
If git is installed and the current directory is part of a git tree,
then also shows all information from \code{\link{prompt_git}}.
}
\seealso{
Other example prompts: 
\code{\link{new_prompt_powerline}()},
\code{\link{prompt_fancy}()},
\code{\link{prompt_git}()},
\code{\link{prompt_mem}()},
\code{\link{prompt_runtime}()},
\code{\link{prompt_status}()}
}
\concept{example prompts}
