% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{prompt_dir}
\alias{prompt_dir}
\title{Choose a directory}
\usage{
prompt_dir(prompt = "", ...)
}
\arguments{
\item{prompt}{String. Prompt to guide user.}
}
\description{
Prompt user to choose a directory.
}
\details{
\describe{
  \item{Mac}{Applescript is used to prompt to open a system file browser.
  The user is restricted to selecting only directories.}
  \item{Windows}{The Windows only \code{utils::choose.dir}}
  \item{Unix}{If \code{tcltk} is available that is used. Otherwise, a text
  prompt is generated via \link[base]{readline}}
  \item{unknown}{In the case of an unknown system, a text prompt is generated
  via \link[base]{readline}}
}
}
\examples{
\dontrun{
  prompt_dir('Please select a directory')
}
}

