% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_fancy.R
\name{prompt_fancy}
\alias{prompt_fancy}
\title{A fancy prompt, showing probably too much information}
\usage{
prompt_fancy(expr, value, ok, visible)
}
\arguments{
\item{expr}{Evaluated expression.}

\item{value}{Its value.}

\item{ok}{Whether the evaluation succeeded.}

\item{visible}{Whether the result is visible.}
}
\description{
It also uses color, on terminals that support it.
Is shows: \itemize{
  \item Status of last command.
  \item Memory usage of the R process.
  \item Package being developed using devtools, if any.
  \item Git branch and state of the working tree if within a git tree.
}
}
\seealso{
Other example prompts: \code{\link{prompt_devtools}},
  \code{\link{prompt_error}}, \code{\link{prompt_git}},
  \code{\link{prompt_memuse}}, \code{\link{prompt_runtime}}
}
\concept{example prompts}
