% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{prompt_file}
\alias{prompt_file}
\title{Choose a single file}
\usage{
prompt_file(prompt = "", ...)
}
\arguments{
\item{prompt}{String. Message to guide user.}

\item{...}{Further arguments passed to \link[base]{file.choose}}
}
\description{
Prompt a user to choose a single file.
}
\details{
This is an alias for \link[base]{file.choose} that messages the
user with a prompt to guide them on what they should be choosing.
}
\examples{
\dontrun{
  prompt_file('Please select a file')
}
}

