% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{prompt_files}
\alias{prompt_files}
\title{Choose multiple files}
\usage{
prompt_files(prompt = "", ...)
}
\arguments{
\item{prompt}{String. Prompt to guide user.}
}
\description{
Prompt a user to choose one or more files.
}
\details{
\describe{
  \item{mac}{Implementation details}
  \item{windows}{Implementation details}
  \item{unix}{Implementation details}
  \item{unknown}{Implementation details}
}
}
\examples{
\dontrun{
  prompt_files('Please select one or more files')
}

}

