% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_git.R
\name{prompt_git}
\alias{prompt_git}
\title{A git prompt, showing some information about the current git state}
\usage{
prompt_git(...)
}
\description{
Is shows: \itemize{
  \item Git branch and state of the working tree if within a git tree.
}
}
\seealso{
Other prompts: \code{\link{prompt_fancy}},
  \code{\link{prompt_memuse}}
}
\concept{prompts}
