% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_memuse.R
\name{prompt_memuse}
\alias{prompt_memuse}
\title{Prompt that shows the current memory usage of the R process}
\usage{
prompt_memuse(...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
Prompt that shows the current memory usage of the R process
}
\seealso{
Other prompts: \code{\link{prompt_fancy}}
}
\concept{prompts}
