% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt-runtime.R
\name{prompt_runtime}
\alias{prompt_runtime}
\title{A prompt that shows the CPU time used by the last top level expression}
\usage{
prompt_runtime(...)
}
\arguments{
\item{...}{Arguments, ignored.}
}
\description{
A prompt that shows the CPU time used by the last top level expression
}
\seealso{
Other example prompts: 
\code{\link{new_prompt_powerline}()},
\code{\link{prompt_devtools}()},
\code{\link{prompt_fancy}()},
\code{\link{prompt_git}()},
\code{\link{prompt_mem}()},
\code{\link{prompt_status}()}
}
\concept{example prompts}
