% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt-status.R
\name{prompt_status}
\alias{prompt_status}
\title{A prompt that shows the status (OK or error) of the last expression}
\usage{
prompt_status(expr, value, ok, visible)
}
\arguments{
\item{expr}{Evaluated expression.}

\item{value}{Its value.}

\item{ok}{Whether the evaluation succeeded.}

\item{visible}{Whether the result is visible.}
}
\value{
\code{prompt_status()} returns the prompt string.
}
\description{
A prompt that shows the status (OK or error) of the last expression
}
\seealso{
Other example prompts: 
\code{\link{new_prompt_powerline}()},
\code{\link{prompt_devtools}()},
\code{\link{prompt_fancy}()},
\code{\link{prompt_git}()},
\code{\link{prompt_mem}()},
\code{\link{prompt_runtime}()}
}
\concept{example prompts}
