% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_prompt.R
\name{set_prompt}
\alias{set_prompt}
\alias{set_prompt}
\alias{reset_prompt}
\title{Set the dynamic prompt}
\usage{
set_prompt(fun)

reset_prompt()
}
\arguments{
\item{fun}{whatever \code{\link[rlang]{as_function}} can turn into a function with no parameters}
}
\description{
Set the dynamic prompt
}
\examples{
set_prompt( ~"R> " )
set_prompt( ~paste( format(Sys.time(), "\%H:\%M:\%S"), " > " ) )

}
