% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt-powerline.R
\name{make_prompt_powerline}
\alias{make_prompt_powerline}
\title{This is a Powerline-like prompt}
\usage{
make_prompt_powerline(
  col_mem = "#404040",
  col_path = "#0055d4",
  col_pkg = "gold4",
  col_git = "#7d26cd"
)
}
\arguments{
\item{col_mem, col_path, col_pkg, col_git}{Background colors for the
sections: allocated memory, working directory, devtools loaded
package(s), git info.}
}
\value{
`make_prompt_powerline()` returns a function that you can
use with [set_prompt()].
}
\description{
It is inspired by the <https://github.com/powerline/powerline>
project. This prompt uses some Unicode glyphs that work best
with the fonts specifically modified for Powerline:
https://github.com/powerline/fonts
}
