% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{prompt_devtools}
\alias{prompt_devtools}
\alias{devtools_packages}
\title{Example prompt that shows the package being developed with devtools}
\usage{
prompt_devtools(...)

devtools_packages()
}
\arguments{
\item{...}{Ignored.}
}
\description{
If git is installed and the current directory is part of a git tree,
then also shows all information from \code{\link{prompt_git}}.
}
\examples{
prompt_devtools()
}
\seealso{
Other example prompts: 
\code{\link{prompt_error}()},
\code{\link{prompt_fancy}()},
\code{\link{prompt_git}()},
\code{\link{prompt_memuse}()},
\code{\link{prompt_runtime}()}
}
\concept{example prompts}
