% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{prompt_error}
\alias{prompt_error}
\title{A prompt that shows the status (OK or error) of the last expression}
\usage{
prompt_error(expr, value, ok, visible)
}
\arguments{
\item{expr}{Evaluated expression.}

\item{value}{Its value.}

\item{ok}{Whether the evaluation succeeded.}

\item{visible}{Whether the result is visible.}
}
\description{
A prompt that shows the status (OK or error) of the last expression
}
\seealso{
Other example prompts: \code{\link{prompt_devtools}},
  \code{\link{prompt_fancy}}, \code{\link{prompt_git}},
  \code{\link{prompt_memuse}}, \code{\link{prompt_runtime}}
}
\concept{example prompts}
