% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt-git.R
\name{prompt_git}
\alias{prompt_git}
\alias{git_branch}
\alias{git_arrows}
\alias{git_dirty}
\title{An example 'git' prompt}
\usage{
prompt_git(...)

git_branch()

git_arrows()

git_dirty()
}
\arguments{
\item{...}{Unused.}
}
\description{
It shows the current branch, whether there are
commits to push or pull to the default remote,
and whether the working directory is dirty.
}
\examples{
\dontrun{
  set_prompt(prompt_git)
}
}
\seealso{
Other example.prompts: \code{\link{prompt_devtools}},
  \code{\link{prompt_error}}, \code{\link{prompt_fancy}},
  \code{\link{prompt_memuse}}, \code{\link{prompt_runtime}}
}

