% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt-mem.R
\name{prompt_mem}
\alias{prompt_mem}
\alias{memory_usage}
\title{Example prompt that shows the current memory usage of the R process}
\usage{
prompt_mem(...)

memory_usage()
}
\arguments{
\item{...}{Ignored.}
}
\value{
`prompt_mem()` returns the formatted prompt in a string.

`memory_usage()` returns a list with entries:
  * `bytes`: the number of bytes of memory the current process uses.
    This is the 'Resident Set Size', see [ps::ps_memory_info()].
  * `formatted`: string that formats `bytes` nicely, with the appropriate
    unit.
  * `total`: Total physical memory. See [ps::ps_system_memory()].
  * `avail`: the memory that can be given instantly to processes
     without the system going into swap. See [ps::ps_system_memory()].
  * `percent`: Percentage of memory that is taken. See
     [ps::ps_system_memory()].
}
\description{
Example prompt that shows the current memory usage of the R process
}
\details{
`prompt_mem()` is a simple example prompt that shows the physical memory
allocated by the current process.

`memory_usage()` is a utility function that shows memory information
about the current R process and the system. You can use it to create a
custom prompt.
}
\examples{
\dontshow{if (ps::ps_is_supported()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cat(prompt_mem())
\dontshow{\}) # examplesIf}
\dontshow{if (ps::ps_is_supported()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
memory_usage()
\dontshow{\}) # examplesIf}
}
\seealso{
Other example prompts: 
\code{\link{prompt_devtools}()},
\code{\link{prompt_fancy}()},
\code{\link{prompt_git}()},
\code{\link{prompt_status}()}
}
\concept{example prompts}
