% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{prompt_memuse}
\alias{prompt_memuse}
\title{Example prompt that shows the current memory usage of the R process}
\usage{
prompt_memuse(...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
Example prompt that shows the current memory usage of the R process
}
\seealso{
Other example prompts: \code{\link{prompt_devtools}},
  \code{\link{prompt_error}}, \code{\link{prompt_fancy}},
  \code{\link{prompt_git}}, \code{\link{prompt_runtime}}
}
\concept{example prompts}
