% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{prompt_runtime}
\alias{prompt_runtime}
\title{A prompt that shows the CPU time used by the last top level expression}
\usage{
prompt_runtime(...)
}
\arguments{
\item{...}{Arguments, ignored.}
}
\description{
A prompt that shows the CPU time used by the last top level expression
}
\seealso{
Other example prompts: \code{\link{prompt_devtools}},
  \code{\link{prompt_error}}, \code{\link{prompt_fancy}},
  \code{\link{prompt_git}}, \code{\link{prompt_memuse}}
}
\concept{example prompts}
