% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{set_prompt}
\alias{restore}
\alias{set_prompt}
\alias{suspend}
\alias{toggle}
\title{Set the prompt}
\usage{
set_prompt(value)

suspend()

restore()

toggle()
}
\arguments{
\item{value}{A character string for a static prompt, or
a function that is called after the evaluation every expression
typed at the R prompt. The function should always return a
character scalar.}
}
\description{
Set the prompt
}

