% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{set_prompt}
\alias{set_prompt}
\alias{suspend}
\alias{restore}
\alias{toggle}
\title{Set and control the prompt}
\usage{
set_prompt(value)

suspend()

restore()

toggle()
}
\arguments{
\item{value}{A character string for a static prompt, or
a function that is called after the evaluation every expression
typed at the R prompt. The function should always return a
character scalar.}
}
\description{
Set and control the prompt
}
\details{
Function \code{update_prompt()} is used to replace the default \R
prompt with a custom prompt.   A custom prompt can be disabled
with \code{suspend()} and then re-enable with \code{restore()}.
Function \code{toggle()} toggles between the two.
}
